Added an echo location feature for detecting objects rather than
going with my first thought of the motion sensor with the obstacle
avoidance sensor. The obstacle avoidance sensor works but you have to
be very close, the echo location sensor works from a farther distance.
According to the docs it works up to 3-4 meters but I have not tested it
accurately beyond a few feet. This is okay because I want to use This
within a foot or so and it is working well for that.

The existing code was organized into a lib folder to make it more
manageable. The main.c file was also updated to reflect the new changes.

This closes issue #1.